# mypostbank-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI specification - Yasin",
    "description": "OpenAPI Documentation for MyPostBank",
    "termsOfService": "Terms of Service",
    "contact": {
      "name": "Yasin",
      "email": "yasinjb002@gmail.com"
    },
    "license": {
      "name": "License"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/mypostbank",
      "description": "Local ENV"
    },
    {
      "url": "https://yasin",
      "description": "PROV ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/user/loans/request-loan": {
      "post": {
        "tags": [
          "credit-controller"
        ],
        "summary": "Request a new loan",
        "operationId": "requestLoan",
        "parameters": [
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "rate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "duration",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "accountId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "form",
                  "proof",
                  "statements"
                ],
                "type": "object",
                "properties": {
                  "form": {
                    "type": "string",
                    "format": "binary"
                  },
                  "statements": {
                    "type": "string",
                    "format": "binary"
                  },
                  "proof": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/loans/{creditId}": {
      "get": {
        "tags": [
          "credit-controller"
        ],
        "summary": "Get loan by ID",
        "operationId": "getLoanById",
        "parameters": [
          {
            "name": "creditId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved loan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditDTO"
                }
              }
            }
          },
          "404": {
            "description": "Loan not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditDTO"
                }
              }
            }
          }
        }
      }
    },
    "/user/loans/{creditId}/due-payments": {
      "get": {
        "tags": [
          "credit-controller"
        ],
        "summary": "Get due payments",
        "operationId": "getDuePayments",
        "parameters": [
          {
            "name": "creditId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepaymentDetailDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/loans/{creditId}/all-payments": {
      "get": {
        "tags": [
          "credit-controller"
        ],
        "summary": "Get all payments",
        "operationId": "getAllRepaymentsByUser",
        "parameters": [
          {
            "name": "creditId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepaymentDetailDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/loans/{crediId}/paid-payments": {
      "get": {
        "tags": [
          "credit-controller"
        ],
        "summary": "Get paid payments",
        "operationId": "getPaidPayments",
        "parameters": [
          {
            "name": "crediId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepaymentDetailDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreditDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number"
          },
          "interestRate": {
            "type": "number"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "ONGOING",
              "PENDING",
              "REJECTED",
              "PAID"
            ]
          },
          "account": {
            "type": "integer",
            "format": "int32"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "applicationForm": {
            "type": "string",
            "format": "binary"
          },
          "bankStatements": {
            "type": "string",
            "format": "binary"
          },
          "proofOfIncome": {
            "type": "string",
            "format": "binary"
          },
          "owner": {
            "type": "string"
          }
        }
      },
      "RepaymentDetailDTO": {
        "type": "object",
        "properties": {
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "principalPayment": {
            "type": "number"
          },
          "interestPayment": {
            "type": "number"
          },
          "remainingBalance": {
            "type": "number"
          },
          "paymentAmount": {
            "type": "number"
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "PAID",
              "NOT_PAID",
              "PARTIAL",
              "OVERDUE",
              "CANCELLED",
              "INSUFFICIENT_FUNDS"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}